backend/requirements.txt
# ODMS Backend Dependencies
# Python Flask API Server with OR-Tools Route Optimization

# Web Framework
Flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-jwt-extended==4.6.0
bcrypt==4.2.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1

# Route Optimization
ortools==9.11.4210

# Data Processing (ETL Pipeline)
pandas==2.2.3
numpy==2.2.1

# HTTP requests (for replay engine)
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0

# Development utilities
gunicorn==23.0.0
